hf-bg-removal-service/requirements.txt
Flask>=3.0,<4.0
flask-cors>=5.0,<6.0
gunicorn>=23.0,<24.0
Pillow>=10.0,<12.0
rembg>=2.0,<3.0
onnxruntime>=1.18,<2.0

